+++ NEWS.md
@@ -5,6 +5,7 @@
 ## bug fixes
 
 * fixed problem of updating cache. Didn't handle data.frame object that was thought to be a vector correctly!
+* fixed a bug in the BPF export function, which meant that WAVE files were only copied into one session
 
 # emuR 0.2.3